apps/recommendation/requirements.txt
pandas>=2.0.0
redis>=5.0.0
cassandra-driver>=3.28.0
implicit>=0.7.0
scikit-learn>=1.3.0
networkx>=3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
lightfm>=1.17
celery[redis]>=5.3.0
annoy>=1.17.0
scipy>=1.11.0
numpy>=1.24.0
